Unify util modules and make use of utils in p45 and adsim
