+++ src/dodal/adsim.py
@@ -1,22 +1,18 @@
+import os
 import socket
-
-from pydantic import BaseSettings
 
 from dodal.devices.adsim import SimStage
 from dodal.devices.areadetector import AdSimDetector
 
+from .utils import get_hostname
 
-# Settings can be customized via environment variables
-class Settings(BaseSettings):
-    pv_prefix: str = socket.gethostname().split(".")[0]
-
-
-_settings = Settings()
+# Default prefix to hostname unless overriden with export PREFIX=<prefix>
+PREFIX: str = os.environ.get("PREFIX", get_hostname())
 
 
 def stage(name: str = "sim_motors") -> SimStage:
-    return SimStage(name=name, prefix=f"{_settings.pv_prefix}-MO-SIM-01:")
+    return SimStage(name=name, prefix=f"{PREFIX}-MO-SIM-01:")
 
 
 def det(name: str = "adsim") -> AdSimDetector:
-    return AdSimDetector(name=name, prefix=f"{_settings.pv_prefix}-AD-SIM-01:")
+    return AdSimDetector(name=name, prefix=f"{PREFIX}-AD-SIM-01:")
